requirements.txt
alembic==0.8.3
bcrypt==2.0.0
blinker==1.4
cffi==1.3.1
dominate==2.1.16
Flask==0.10.1
Flask-Bcrypt==0.7.1
Flask-Bootstrap==3.3.5.7
Flask-Login==0.3.2
Flask-Mail==0.9.1
Flask-Markdown==0.3
Flask-Migrate==1.6.0
Flask-MySQL==1.3
Flask-Script==2.0.5
Flask-SQLAlchemy==2.1
Flask-Uploads==0.1.3
Flask-WTF==0.12
gunicorn==19.4.1
itsdangerous==0.24
Jinja2==2.8
Mako==1.0.3
Markdown==2.6.5
MarkupSafe==0.23
MySQL-python==1.2.5
pycountry==1.18
pycparser==2.14
python-editor==0.5
python-slugify==1.1.4
six==1.10.0
SQLAlchemy==1.0.10
Unidecode==0.4.18
visitor==0.1.2
Werkzeug==0.11.2
wheel==0.24.0
WTForms==2.0.2
